+++ src/Voting.Lib.Grpc/Interceptors/RequestLoggerInterceptor.cs
@@ -2,6 +2,7 @@
 // For license information see LICENSE file
 
 using System;
+using System.Text;
 using System.Threading.Tasks;
 using Google.Protobuf;
 using Grpc.Core;
@@ -9,6 +10,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Serilog.Context;
+using Voting.Lib.Grpc.Configuration;
 
 namespace Voting.Lib.Grpc.Interceptors;
 
@@ -20,18 +22,22 @@ public class RequestLoggerInterceptor : RequestInterceptor
     private const string LogTypeProperty = "GrpcRequestLogType";
     private readonly ILogger<RequestLoggerInterceptor> _logger;
     private readonly IWebHostEnvironment _environment;
+    private readonly RequestLoggerInterceptorConfig _config;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="RequestLoggerInterceptor"/> class.
     /// </summary>
     /// <param name="logger">The logger.</param>
     /// <param name="environment">The web host environment.</param>
+    /// <param name="config">The interceptor configuration.</param>
     public RequestLoggerInterceptor(
         ILogger<RequestLoggerInterceptor> logger,
-        IWebHostEnvironment environment)
+        IWebHostEnvironment environment,
+        RequestLoggerInterceptorConfig config)
     {
         _logger = logger;
         _environment = environment;
+        _config = config;
     }
 
     /// <inheritdoc/>
@@ -50,13 +56,33 @@ public class RequestLoggerInterceptor : RequestInterceptor
         using (LogContext.PushProperty(LogTypeProperty, true))
         {
             _logger.LogDebug(
-                "{Uri}{NewLine}{RequestData}{NewLine}",
+                "{RequestPath}{NewLine}{RequestData}{NewLine}{RequestHeaders}{NewLine}",
                 context.Method,
                 Environment.NewLine,
                 protoRequest.ToString(),
+                Environment.NewLine,
+                GetRequestHeaders(context),
                 Environment.NewLine);
         }
 
         return Task.CompletedTask;
+    }
+
+    private string GetRequestHeaders(ServerCallContext context)
+    {
+        var sb = new StringBuilder();
+
+        foreach (var header in _config.RequestHeadersToLog)
+        {
+            var headerValue = context.RequestHeaders.GetValue(header);
+            if (string.IsNullOrEmpty(headerValue))
+            {
+                continue;
+            }
+
+            sb.Append(header).Append(": ").AppendLine(headerValue);
+        }
+
+        return sb.ToString();
     }
 }