+++ src/Voting.Lib.Iam/Models/SecureConnectTokenClaimTypes.cs
@@ -1,0 +1,11 @@
+// (c) Copyright 2022 by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Voting.Lib.Iam.Models;
+
+internal static class SecureConnectTokenClaimTypes
+{
+    internal const string TokenType = "type";
+
+    internal const string Actor = "act";
+}