+++ src/Voting.Lib.Iam/Models/SecureConnectTokenTypes.cs
@@ -1,0 +1,11 @@
+// (c) Copyright 2022 by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Voting.Lib.Iam.Models;
+
+internal static class SecureConnectTokenTypes
+{
+    internal const string AccessToken = "access_token";
+    internal const string OnBehalfOfToken = "ob_token";
+    internal const string RoleToken = "role_token";
+}