+++ src/Voting.Lib.VotingExports/Repository/Ausmittlung/AusmittlungPdfMajorityElectionTemplates.cs
@@ -17,7 +17,7 @@ public static class AusmittlungPdfMajorityElectionTemplates
     public static readonly TemplateModel CountingCircleProtocol = new TemplateModel
     {
         Key = "majority_election_counting_circle_protocol",
-        Filename = "Gemeindeprotokoll - {0}",
+        Filename = "Majorz_Gemeindeprotokoll_{0}_{1}",
         Description = "Gemeindeprotokoll",
         Format = ExportFileFormat.Pdf,
         EntityType = EntityType.MajorityElection,
@@ -31,8 +31,8 @@ public static class AusmittlungPdfMajorityElectionTemplates
     public static readonly TemplateModel EndResultProtocol = new TemplateModel
     {
         Key = "majority_election_end_result_protocol",
-        Filename = "Wahlprotokoll Gesamtergebnis aller Einheiten - {0}",
-        Description = "Gesamtergebnis aller Einheiten",
+        Filename = "Majorz_Wahlprotokoll_{0}_{1}",
+        Description = "Wahlprotokoll",
         Format = ExportFileFormat.Pdf,
         EntityType = EntityType.MajorityElection,
         ResultType = ResultType.PoliticalBusinessResult,
@@ -45,8 +45,22 @@ public static class AusmittlungPdfMajorityElectionTemplates
     public static readonly TemplateModel EndResultDetailProtocol = new TemplateModel
     {
         Key = "majority_election_end_result_detail_protocol",
-        Filename = "Detailergebniss aller einseh. Einheiten - {0}",
-        Description = "Detailergebnisse aller einsehbaren Einheiten",
+        Filename = "Majorz_Detailergebnisse_{0}_{1}",
+        Description = "Detailergebnisse",
+        Format = ExportFileFormat.Pdf,
+        EntityType = EntityType.MajorityElection,
+        ResultType = ResultType.PoliticalBusinessResult,
+        GeneratedBy = VotingApp.VotingAusmittlung,
+    };
+
+    /// <summary>
+    /// Gets the end result detail protocol export template.
+    /// </summary>
+    public static readonly TemplateModel EndResultDetailWithoutEmptyAndInvalidVotesProtocol = new TemplateModel
+    {
+        Key = "majority_election_end_result_detail_without_empty_and_invalid_votes_protocol",
+        Filename = "Majorz_Detailergebnisse_exkl_{0}_{1}",
+        Description = "Detailergebnisse_exkl",
         Format = ExportFileFormat.Pdf,
         EntityType = EntityType.MajorityElection,
         ResultType = ResultType.PoliticalBusinessResult,
@@ -72,6 +86,7 @@ public static class AusmittlungPdfMajorityElectionTemplates
         CountingCircleProtocol,
         EndResultProtocol,
         EndResultDetailProtocol,
+        EndResultDetailWithoutEmptyAndInvalidVotesProtocol,
         ResultBundleReview,
     };
 }